% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist.torus.R
\name{dist.torus}
\alias{dist.torus}
\title{Distance on the two-dimensional torus}
\usage{
dist.torus(x, y)
}
\arguments{
\item{x}{A vector of two coordinates in the two-dimensional flat torus, parameterized as the periodic [0,1) x [0,1).}

\item{y}{A vector of two coordinates in the two-dimensional flat torus, parameterized as the periodic [0,1) x [0,1).}
}
\value{
The distance on the torus between x and y
}
\description{
Distance between two points on the two-dimensional torus (periodic [0,1] x [0,1])
}
\examples{
set.seed(10)
# Uniform distribution on [0,1] x [0,1]
x <- uniformly::runif_in_cube(1, 2, c(0.5, 0.5), 0.5) 
y <- uniformly::runif_in_cube(1, 2, c(0.5, 0.5), 0.5)
dist.torus(x, y)

# Bivariate von Mises distribution
x <- uniformly::runif_in_cube(1, 2, c(0.5, 0.5), 0.5) 
y <- BAMBI::rvmcos(1, kappa1 = 1, kappa2 = 1, mu1 = 0, mu2 = 0)/(2*pi)
dist.torus(x, y)

}
