% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geodesic.projection.R
\name{geodesic.projection}
\alias{geodesic.projection}
\title{Geodesic projection of a pair of samples on the torus}
\usage{
geodesic.projection(u, data_1, data_2, do_plots = FALSE, size_points = 1)
}
\arguments{
\item{u}{A vector of two integers (a,b), with b non-zero.}

\item{data_1}{n x 2 matrix containing n observations in the two-dimensional flat torus, parameterized as the periodic [0,1) x [0,1).}

\item{data_2}{m x 2 matrix containing m observations in the two-dimensional flat torus, parameterized as the periodic [0,1) x [0,1).}

\item{do_plots}{Whether to produce plots illustrating the steps of the algorithm for the given geodesic and samples.}

\item{size_points}{If do_plots is TRUE, the size of the points to plot (to choose according to n,m).}
}
\value{
\itemize{
  \item plot_list - If do_plots is TRUE, a list with the corresponding plots.
  \item circle_one - The projected sample data_1 parameterized on the circle (periodic [0,1)).
  \item circle_two - The projected sample data_2 parameterized on the circle (periodic [0,1)).
  \item proj_data - A data frame containing the coordinates on [0,1]x[0,1] of both samples and the corresponding projections.

}
}
\description{
Given a vector of integers u=(a,b), two samples on the torus are projected to the
closed geodesic given by the canonical projection of the straight line with director
vector u and containing the origin.
}
\examples{

data_1 <- BAMBI::rvmcos(30) / (2 * pi)
data_2 <- BAMBI::rvmcos(30) / (2 * pi)
pr <- geodesic.projection(u = c(-1, -2), data_1, data_2, do_plots = TRUE)
pr$plot_list[[1]]
pr$plot_list[[2]]
pr$plot_list[[3]]
pr$plot_list[[4]]
pr$plot_list[[5]]
pr$circle_one; pr$circle_two

}
