% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgeodesic.R
\name{rgeodesic}
\alias{rgeodesic}
\title{Closed geodesics sampling}
\usage{
rgeodesic(ng, p = 0.1)
}
\arguments{
\item{ng}{The size of the geodesics sample.}

\item{p}{The parameter p of the geometric distribution of B.}
}
\value{
A ng x 2 matrix containing the vectors corresponding to the sampled closed geodesics.
}
\description{
Samples a number of different closed geodesics on the torus. Each geodesic is represented by a vector of integers (B, A),
where B follows a geometric distribution of parameter p and, for a given B=b, A is uniform in (0,1,...,b).
}
\details{
For big sample sizes, the value of p must be reduced in order to get a sample of different
geodesics. This yields to less simple geodesics, with more revolutions over the torus.
}
\examples{

rgeodesic(5, p = 0.1)

# For bigger sample sizes, p must be reduced.
rgeodesic(100, p = 0.01)

}
\references{
[1] Petroni, N. C.(2019). Taking rational numbers at random. arXiv:1908.06944v1.
}
