% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.null.stat.R
\name{sim.null.stat}
\alias{sim.null.stat}
\title{Wasserstein statistic null distribution}
\usage{
sim.null.stat(NR, NC = 1, n = 30)
}
\arguments{
\item{NR}{Number of replicas to simulate}

\item{NC}{Number of cores for parallel computation}

\item{n}{Sample size for the simulated samples}
}
\value{
A sample of size NR simulating the null distribution of the statistic.
}
\description{
Simulates Wasserstein statistic null distribution
}
\examples{
sim.null.stat(100)

}
\references{
[1] Delon, J., Salomon, J., Sobolevskii, A.: Fast transport optimization for Monge costs on the circle. SIAM J. Appl. Math. 70(7), 2239–2258 (2010).
[2] RAMDAS, A., GARCIA, N. and CUTURI, M. (2015). On Wasserstein Two Sample Testing and Related Families of Nonparametric Tests. Entropy 19.
}
