% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.s1.R
\name{stat.s1}
\alias{stat.s1}
\title{Wasserstein distribution-free statistic on S^1}
\usage{
stat.s1(x, y)
}
\arguments{
\item{x}{numerical vector of sample in [0,1).}

\item{y}{numerical vector of sample in [0,1).}
}
\value{
The statistic realization of samples x, y.
}
\references{[1] Delon, J., Salomon, J., Sobolevskii, A.: Fast transport optimization for Monge costs on the circle. SIAM J. Appl. Math. 70(7), 2239–2258 (2010).
[2] RAMDAS, A., GARCIA, N. and CUTURI, M. (2015). On Wasserstein Two Sample Testing and Related Families of Nonparametric Tests. Entropy 19.
}
\description{
Computes Wasserstein statistic between two samples on S^1, parameterized as a periodic [0,1).
}
\examples{
set.seed(10)
stat.s1(runif(50), runif(50))
stat.s1(runif(50), as.numeric(circular::rvonmises(50, pi, 1)/(2*pi)))
}
