% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twosample.geodesic.torus.test.R
\name{twosample.geodesic.torus.test}
\alias{twosample.geodesic.torus.test}
\title{Geodesic two-sample goodness-of-fit on the torus}
\usage{
twosample.geodesic.torus.test(
  sample_1,
  sample_2,
  n_geodesics = 1,
  NC_geodesic = 1,
  geodesic_list = NULL,
  sim_null = NULL,
  NR = 500,
  NC = 1,
  n = 30
)
}
\arguments{
\item{sample_1}{n x 2 matrix containing n observations in the two-dimensional flat torus, parameterized as the periodic [0,1) x [0,1).}

\item{sample_2}{n x 2 matrix containing n observations in the two-dimensional flat torus, parameterized as the periodic [0,1) x [0,1).}

\item{n_geodesics}{The number of closed geodesics where samples must be projected.}

\item{NC_geodesic}{The number of cores for computing parallelly the circular tests.}

\item{geodesic_list}{If NULL, geodesics are chosen randomly with the function rgeodesic.
Else, a list of vectors containing the director vectors of the chosen geodesics.}

\item{sim_null}{The simulated null distribution of the circle test statistic. It NULL, the null distribution is simulated with the given parameters (very time consuming).}

\item{NR}{The number of replicas if simulation is required.}

\item{NC}{The number of cores if parallel simulation for sim_null is required.}

\item{n}{The sample sizes of the simulated samples is simulation is required.}
}
\value{
The p-value for the two-sample test on the torus.
}
\description{
Performs a two-sample goodness-of-fit test for measures supported on the torus, by testing
the equality of their projected distributions on ng closed geodesics and combining the results into a global p-value.
}
\examples{

n <- 100 # Sample size

# Simulate the null distribution of the circle test statistic

sim_free_null <- sim.null.stat(500, NC = 2)

# Bivariate von Mises distributions
samp_1 <- BAMBI::rvmcos(n) / (2 * pi) 
samp_2 <- BAMBI::rvmcos(n) / (2 * pi)

#4 geodesics are chosen randomly
twosample.geodesic.torus.test(samp_1, samp_2, n_geodesics = 3,
NC_geodesic = 2, sim_null = sim_free_null) 

#4 geodesics are chosen a priori
glist <- list(c(1, 0), c(0, 1), c(1, 1), c(2, 3))
twosample.geodesic.torus.test(samp_1, samp_2, geodesic_list = glist, 
NC_geodesic = 2, sim_null = sim_free_null) 


}
