% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twosample.test.s1.R
\name{twosample.test.s1}
\alias{twosample.test.s1}
\title{Two-sample goodness-of-fit on the circle}
\usage{
twosample.test.s1(x, y, sim_null = NULL, NR = 500, NC = 1, n = 30)
}
\arguments{
\item{x}{numerical vector of sample in [0,1).}

\item{y}{numerical vector of sample in [0,1).}

\item{sim_null}{The simulated null distribution of the statistic. If NULL, the distribution is simulated with the given parameters (very time consuming).}

\item{NR}{The number of replicas if simulation is required.}

\item{NC}{The number of cores if parallel simulation is required.}

\item{n}{The sample sizes of the simulated samples is simulation is required.}
}
\value{
\itemize{
  \item stat - The test statistic.
  \item pvalue - The test p-value.
}
}
\description{
Performs a two-sample goodness-of-fit test for measures supported on the circle, based on a
distribution-free Wasserstein statistic.
}
\examples{

n <- 50 # Sample size

# Simulate the statistic null distribution
NR <- 100
sim_free_null <- sim.null.stat(500, NC = 2)

x <- runif(n, 0, 1)
y <- runif(n, 0, 1)
twosample.test.s1(x, y, sim_free_null) 

x <- as.numeric(circular::rvonmises(n, pi, 1)/(2*pi))
y <- as.numeric(circular::rvonmises(n, pi, 0)/(2*pi))
twosample.test.s1(x, y, sim_free_null) 

}
\references{
[1] Delon, J., Salomon, J., Sobolevskii, A.: Fast transport optimization for Monge costs on the circle. SIAM J. Appl. Math. 70(7), 2239–2258 (2010).
[2] RAMDAS, A., GARCIA, N. and CUTURI, M. (2015). On Wasserstein Two Sample Testing and Related Families of Nonparametric Tests. Entropy 19.
}
