% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twosample.ubound.torus.test.R
\name{twosample.ubound.torus.test}
\alias{twosample.ubound.torus.test}
\title{Asymptotic two-sample goodness-of-fit on the torus}
\usage{
twosample.ubound.torus.test(sample_1, sample_2)
}
\arguments{
\item{sample_1}{n x 2 matrix containing n observations in the two-dimensional flat torus, parametrized as the periodic [0,1) x [0,1).}

\item{sample_2}{n x 2 matrix containing n observations in the two-dimensional flat torus, parametrized as the periodic [0,1) x [0,1).}
}
\value{
The p-value for the two-sample test on the torus.
}
\description{
Performs a two-sample goodness-of-fit test for measures supported on the torus, using a p-value upper bound. The
test is asymptotically consistent at level alpha.
}
\examples{

n <- 2000 # Sample size

set.seed(10)
 # Bivariate von Mises distribution
samp_1 <- BAMBI::rvmcos(n, kappa1 = 1, kappa2 = 1, mu1 = 0, mu2 = 0)/(2*pi)
samp_2 <- BAMBI::rvmcos(n, kappa1 = 1, kappa2 = 1, mu1 = 0, mu2 = 0)/(2*pi)
twosample.ubound.torus.test(samp_1, samp_2) 

samp_1 <- BAMBI::rvmcos(n ,kappa1 = 0, kappa2 = 0, mu1 = 0.5, mu2 = 0.5)/(2*pi)
samp_2 <- BAMBI::rvmcos(n, kappa1 = 1, kappa2 = 1, mu1 = 0.5, mu2 = 0.5)/(2*pi)
twosample.ubound.torus.test(samp_1, samp_2) 

}
